Fury works with all Graph and point types!
No extra serializers need to be registered, since the logic was pushed into Graph's implementation of Externalizable. And Graph is the only class that needed it at all! It was only required because a NodeMap is present in each Graph, but not usable for ser/deser. The Externalizable code writes the vertices only, no Nodes.